refactor: 불필요한 메소드 삭제
